an initial idea about how we might manage a list of examples
Thought adding metadata to each example would make this easier
to maintain than if we had to separately add things to a list.
Happy to have a go at writing something to compile the list of
examples. Now that I write this, I seem to remember there is a
plugin - mkdocs-gen-files?